requirements.txt
altair
numpy
pandas
pydeck
streamlit
pyenchant